BIGUPDATE [8.2]

+++ application/cron/manifest.php
@@ -2,10 +2,10 @@
 class Manifest {
 	
 	function __construct() {
-		$file = PATH_UPLOADS."manifest".DS;
+		$file = PATH_MANIFEST;
 		$list = read_dir($file);
 		for($i=0;$i<sizeof($list);$i++) {
-			if($list[$i]!=".htaccess" || $list[$i]!="index.html" || $list[$i]!="index.php") {
+			if($list[$i]!=".htaccess" && $list[$i]!="index.html" && $list[$i]!="index.php") {
 				unlink($file.$list[$i]);
 			}
 		}
